Clean up color dictionary view helpers
