+++ frontend/index.html
@@ -105,10 +105,8 @@
     <script src="js/utils/formula-utils.js"></script>
     <!-- Message utility for unified notifications -->
     <script src="js/utils/message.js"></script>
-    <!-- Color converter utility for RGB/CMYK/HEX conversion -->
-    <script src="js/utils/color-converter.js"></script>
-    <!-- Color conversion utilities for HSL/LAB -->
-    <script src="js/utils/colorConversion.js"></script>
+    <!-- Unified color utility toolkit (RGB/HSL/CMYK/LAB conversions + extraction) -->
+    <script src="js/utils/color-tools.js"></script>
     <!-- Color distance calculations (Delta E) -->
     <script src="js/utils/colorDistance.js"></script>
     <!-- Color mapping utilities for grid and wheel -->
@@ -134,6 +132,12 @@
     <script src="js/components/common/ConflictResolver.js"></script>
     <!-- 通用操作按钮组件 - 暂时禁用 -->
     <!-- <script src="js/components/common/ActionButton.js"></script> -->
+    <!-- 自配色组件 mixins -->
+    <script src="js/components/custom-colors/mixins/filtering.js"></script>
+    <script src="js/components/custom-colors/mixins/pagination.js"></script>
+    <script src="js/components/custom-colors/mixins/selection.js"></script>
+    <script src="js/components/custom-colors/mixins/media.js"></script>
+    <script src="js/components/custom-colors/mixins/duplicates.js"></script>
     <!-- 定义全局变量: CustomColorsComponent -->
     <script src="js/components/custom-colors.js"></script>
     
@@ -146,6 +150,13 @@
     <script src="js/components/category-manager.js"></script>
     <!-- 高级颜色选择对话框组件 -->
     <script src="js/components/color-palette-dialog.js"></script>
+    <!-- 自配色字典模块化依赖 -->
+    <script src="js/components/color-dictionary/helpers.js"></script>
+    <script src="js/components/color-dictionary/mixins/data.js"></script>
+    <script src="js/components/color-dictionary/mixins/interactions.js"></script>
+    <script src="js/components/color-dictionary/views/list-view.js"></script>
+    <script src="js/components/color-dictionary/views/hsl-view.js"></script>
+    <script src="js/components/color-dictionary/views/wheel-view.js"></script>
     <!-- 自配色字典页面组件 -->
     <script src="js/components/color-dictionary.js"></script>
     <!-- 计算器浮层组件 (步骤3) -->
